% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_run.R
\name{canonicalModel_}
\alias{canonicalModel_}
\title{Decifer a 'model parameter' and run appropriate \code{glm_...} function.}
\usage{
canonicalModel_(model)
}
\arguments{
\item{model}{string, function or object representing a model type.}
}
\value{
model type object
}
\description{
Return canonical model from model type string, function of object. Returns
a model as returned by \code{glm_gaussian()} and others, based on a string,
function or model type object (i.e. \code{"glm_gaussian"}, \code{glm_gaussian} or
\code{glm_gaussian()}).
}
