% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combiner.R
\name{combiner}
\alias{combiner}
\title{combiner combines the miRNA and mRNA files}
\usage{
combiner(mRNA, miRNA, miRNA_select)
}
\arguments{
\item{mRNA}{Matrix or data.frame mRNA/RNA from transforemed diff expression file (genrated
using TZtranz)}

\item{miRNA}{Matrix or data frame miRNA from transforemed diff file (genrated
using TZtranz)}

\item{miRNA_select}{A vector of character's for miRNAs which the user is interested in investigating if glm is use 1 miRNA
should be input. If multivariate several miRNAs should be imported, same goes for interaction determination for miRNAs. Note
we do not recommend more than 3-4 miRNAs at a time for the latter cases.}
}
\value{
A dataframe which includes only mRNAs and miRNA intersection for the next estimation geneVari output.
}
\description{
This function makes and intersection dataframe for mRNA and miRNA/s of interest to be tested.
}
\examples{
miRNA_select <- c("ebv-mir-bart9-5p")
x <- combiner(mRNA, miRNA, miRNA_select)
}
\keyword{miRNA_mRNA_Intersection}
