% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corMirnaRna.R
\name{corMirnaRna}
\alias{corMirnaRna}
\title{corMirnaRna correlation for miRNA and mRNA}
\usage{
corMirnaRna(mRNA, miRNA, method = "pearson")
}
\arguments{
\item{mRNA}{mRNA file generated from foldchanges (FC) obj of the one2OneRnaMiRNA}

\item{miRNA}{miRNA file generated from foldchanges (FC) obj of the one2OneRnaMiRNA}

\item{method}{Default is "pearson" else use "kendall" or "spearman"}
}
\value{
Correlation data.frame
}
\description{
This function uses the output of one2OneRnaMiRNA and returns the correlation dataframe
}
\examples{
x <- corMirnaRna(mRNA_fc, miRNA_fc, method = "spearman")
}
\keyword{Correlation}
\keyword{Threshold}
\keyword{miRanda}
\keyword{miRanda,}
\keyword{with}
