% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{corr_0}
\alias{corr_0}
\title{This is data is the mRNA FC and miRNA FC correlation data.
This data set is used in documentation examples.}
\description{
This is data is the mRNA FC and miRNA FC correlation data.
This data set is used in documentation examples.
}
\keyword{data}
