% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawInterPlots.R
\name{drawInterPlots}
\alias{drawInterPlots}
\title{drawInterPlots for finInterResult miRNA and mRNA Interrelation real data}
\usage{
drawInterPlots(mrna, mirna, final_results)
}
\arguments{
\item{mrna}{mRNA results of twoTimePoint function.}

\item{mirna}{miRNA results of twoTimePoint function.}

\item{final_results}{finInterResult miRNA and mRNA interrelation in two timepoints  results in a dataframe.}
}
\value{
par plots
}
\description{
This function draws miRNA, mRNA density plots for miRNA and mRNA Interrelation while
comparing in addition to overall FC_miRNA and FC_mRNA plots from the finInterResult dataframe function.
}
\examples{
x <- drawInterPlots(mRNA_fc2, miRNA_fc2, final_results)
}
\keyword{plot}
