% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{final_results}
\alias{final_results}
\title{This is data is the mRNA FC and miRNA FC correlation/interaction data results after filteration. 
This data set is used in documentation examples.}
\description{
This is data is the mRNA FC and miRNA FC correlation/interaction data results after filteration. 
This data set is used in documentation examples.
}
\keyword{data}
