% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finInterResult.R
\name{finInterResult}
\alias{finInterResult}
\title{finInterResult miRNA and mRNA interrelation in two-time points results in a dataframe.}
\usage{
finInterResult(results)
}
\arguments{
\item{results}{Results from mirandaIntersectInter}
}
\value{
miRNA mRNA interelation dataframe
}
\description{
This function uses the output of one2OneRnaMiRNA and returns a sampled from orig file
interrelation dataframe depending on user sampling selection.
}
\examples{
x <- finInterResult(results)
}
\keyword{Results}
\keyword{dataframe}
