% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneVari.R
\name{geneVari}
\alias{geneVari}
\title{geneVari Makes a list of gene names to be used in the runModels function}
\usage{
geneVari(Combined, miRNA_select)
}
\arguments{
\item{Combined}{the combined file for mRNA and selected miRNAs output of combiner function}

\item{miRNA_select}{The vector of selected miRNA/s}
}
\value{
A vector of characters with defined mRNA dimensions
}
\description{
This function defines the boudnaries of mRNA vs miRNAs of interest to be analysed by the runModels function
}
\examples{
x <- geneVari(Combine, "ebv-mir-bart9-5p")
}
\keyword{dimensions}
