% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geneVariant}
\alias{geneVariant}
\title{This is data is the mRNA expression across samples and miRNA expression data which is to be 
investigated giving directions on which data is miRNA and which is mRNA.
This data set is used in documentation examples.}
\description{
This is data is the mRNA expression across samples and miRNA expression data which is to be 
investigated giving directions on which data is miRNA and which is mRNA.
This data set is used in documentation examples.
}
\keyword{data}
