% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_gaussian.R
\name{glm_gaussian}
\alias{glm_gaussian}
\title{Model functions for GLM with Gaussian model.}
\usage{
glm_gaussian()
}
\value{
structure containing functions \code{fit}, \code{coefficients},
        \code{aic}, \code{data}, \code{pterm}, \code{pmodel}, and a
        character string "glm_gaussian" in \code{model}.
}
\description{
Implements standardaized functions to fit the glm with
Gaussian family and to obtain coefficients, pvalues, etc.
}
\examples{
x <- glm_gaussian()
}
