% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_multi.R
\name{glm_multi}
\alias{glm_multi}
\title{Model functions for GLM with negative binomial family.}
\usage{
glm_multi(
  models = c(glm_gaussian, glm_nb, glm_poisson, glm_zeroinfl_poisson,
    glm_zeroinfl_negbin)
)
}
\arguments{
\item{models}{Model type, one or more of glm_gaussian, glm_nb, glm_poisson, glm_zeroinfl_poisson or glm_zeroinfl_negbin}
}
\value{
structure containing functions \code{fit}, \code{coefficients},
        \code{aic}, \code{data}, \code{pterm}, \code{pmodel}, and a
        character string "glm_multi" in \code{model}.
}
\description{
Runs models 'glm_gaussian', 'glm_nb', 'glm_poisson',
'glm_zeroinfl(poisson)', 'glm_zeroinfl(negbin)' and returns
mode with lowest AIC.
}
\examples{
x <- glm_multi()
}
