% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_nb.R
\name{glm_nb}
\alias{glm_nb}
\title{Model functions for GLM with negative binomial family.}
\usage{
glm_nb()
}
\value{
structure containing functions \code{fit}, \code{coefficients},
        \code{aic}, \code{data}, \code{pterm}, \code{pmodel}, and a
        character string "glm_nb" in \code{model}.
}
\description{
Implements standardaized functions to fit the negative
binomial GLM and to obtain coefficients, pvalues, etc.
}
\examples{
x <- glm_nb()
}
