% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_poisson.R
\name{glm_poisson}
\alias{glm_poisson}
\title{Model functions for GLM with Poisson model.}
\usage{
glm_poisson()
}
\value{
structure containing functions \code{fit}, \code{coefficients},
        \code{aic}, \code{data}, \code{pterm}, \code{pmodel}, and a
        character string "glm_poisson" in \code{model}.
}
\description{
Implements standardaized functions to fit the glm with
Poisson family and to obtain coefficients, pvalues, etc.
}
\examples{
x <- glm_poisson()
}
