% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_zeroinfl.R
\name{glm_zeroinfl_poisson}
\alias{glm_zeroinfl_poisson}
\title{alias for glm_zeroinfl("poisson")}
\usage{
glm_zeroinfl_poisson(...)
}
\description{
alias for glm_zeroinfl("poisson")
}
\examples{
x <- glm_zeroinfl_poisson()
}
