% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importMirandaFile.R
\name{importMirandaFile}
\alias{importMirandaFile}
\title{importMirandaFile Read internal Miranda file}
\usage{
importMirandaFile(fn)
}
\arguments{
\item{fn}{filename}
}
\value{
data.frame containing Miranda data
}
\description{
Reads internal Miranda file from extdata and returns it as a data.frame
}
\examples{
\donttest{
x <- importMirandaFile("Mouse_miRanda.txt")
}
}
