% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFormulaRightSide.R
\name{makeFormulaRightSide}
\alias{makeFormulaRightSide}
\title{makeFormulaRightSide makes right hand side of formula for model variables: vector of indep. variables}
\usage{
makeFormulaRightSide(variables, mode = "multi")
}
\arguments{
\item{variables}{The vector created by miRNA_select}

\item{mode}{Null versus "multi: versus "inter"}

\item{zeroin}{if zero inflated user TRUE default FALSE}
}
\value{
data.frame containing Miranda data
}
\description{
This function make right hand side of formula for model variables: vector of indep. variables (i.e. miRNAs)
mode: 'multi' for simple, 'inter' for model with interactions returns a string in the form "~ a + b", or "~ a + b + a * b"
}
\examples{
\donttest{
x <- makeFormulaRightSide(variables, mode = "multi")
}

}
