% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miRanComp.R
\name{miRanComp}
\alias{miRanComp}
\title{miRanComp comparison of mRNAs present with miRanda file targets}
\usage{
miRanComp(miRNA, miRanda)
}
\arguments{
\item{miRNA}{Matrix or data.frame miRNA/RNA file or transformed diff expression file (generated
using TZtranz)}

\item{miRanda}{A dataframe of miRanda file with miRNA$V1 and miRNA targets miRNA$V2}
}
\value{
An miRNA expression dataframe which includes only Genes/Targets present in miRanda file
}
\description{
This function generates a dataframe consisting of mRNA or miRNAs present in
 miRanda generated file using the miRTarRNASeq:::getInputSpecies() function
}
\examples{
x <- miRanComp(miRNA, miRanda)
}
\keyword{Intersection}
\keyword{expression}
\keyword{miRanda}
