% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sponge.R
\name{miranda_sponge_predict}
\alias{miranda_sponge_predict}
\title{Transform miRanda data for relevant mRNA and miRNA to matrix form compatible with sponge}
\usage{
miranda_sponge_predict(mirna_exp, diff_exp, miranda_data)
}
\arguments{
\item{mirna_exp}{miRna expression data.frame with miRNA for rows and samples for columns}

\item{diff_exp}{mRNA expression data.frame with mRNA for rows and samples for columns}

\item{miranda_data}{miRanda data.frame with the first two columns having miRNA and mRNA names}
}
\value{
matrix adjacency matrix with column names miRNA and row names mRNA
}
\description{
Transforms miRanda data into adjacency matrix, with 1 indicating
presence of a relationship between a mRNA and miRNA, and 0
otherwise. miRanda input is filtered by miRNA and mRNA present in
`mirna_exp` and `diff_expr` row names, respectively.
}
\keyword{SPONGE}
\keyword{ceRNA,}
\keyword{correlation,}
\keyword{miRanda,}
\keyword{partial}
\keyword{sparse}
