% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miRandaIntersect.R
\name{miRandaIntersect}
\alias{miRandaIntersect}
\title{miRandaIntersect Looks for Intersection of Significant output results with miRanda Results from getInputSpeciesDF
function}
\usage{
miRandaIntersect(sig_corrs, corrS, mRNA, miRNA, getInputSpeciesDF)
}
\arguments{
\item{sig_corrs}{correlation matrix, produced by threshSig.}

\item{corrS}{vector of correlations/differences, from the sampCorRnaMirna function.}

\item{getInputSpeciesDF}{miranda data, produced by getInputSpecies.}
}
\value{
An object containing data.frames of significant mRNA, miRNA
            and correlation matrix filtered by miRanda input.
}
\description{
Compares and looks for intersection if significant output results with miRanda Results from getInputSpeciesDF and outputs a final
filterd ourput for only those pairs of miRNA and mRNA which have actually been predicted to be targets
in miRanda file
function
}
\examples{
x <- miRandaIntersect(sig_InterR, outs2, mRNA_fc, miRNA_fc, miRandaM)
}
\keyword{Signficance,}
\keyword{Threshold,}
\keyword{intersect}
