% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirandaIntersectInter.R
\name{mirandaIntersectInter}
\alias{mirandaIntersectInter}
\title{mirandaIntersectInter Looks for Intersection of Significant output results with miRanda Results from getInputSpeciesDF
function}
\usage{
mirandaIntersectInter(sig_corrs, corrS, mRNA, miRNA, getInputSpeciesDF)
}
\arguments{
\item{sig_corrs}{correlation matrix, produced by threshSig}

\item{corrS}{vector of Differences/Correlations, from the sampCorRnaMirna function.}

\item{mRNA}{mRNA FC matrix.}

\item{miRNA}{miRNA FC matrix.}

\item{getInputSpeciesDF}{miranda data, produced by getInputSpecies.}
}
\value{
An object containing data.frames of significant mRNA, miRNA
            and correlation matrix filtered by miranda input.
}
\description{
Compares and looks for intersection if significant output results with miRanda Results from getInputSpeciesDF and outputs a final
filterd ourput for only those pairs of miRNA and mRNA which have actually been predicted to be targets
in miRanda file
function
}
\examples{
x <- mirandaIntersectInter(sig_InterR, outs2, mRNA_fc2, miRNA_fc2, miRandaM)
}
\keyword{Threshold}
\keyword{intersect}
\keyword{miRanda}
\keyword{mirandaIntersectInter}
