% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{miRandaM}
\alias{miRandaM}
\title{This is data is the results file from mouse miRanda getInputSpecies function.
This data set is used in documentation examples.}
\description{
This is data is the results file from mouse miRanda getInputSpecies function.
This data set is used in documentation examples.
}
\keyword{data}
