% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{miRNA2_5}
\alias{miRNA2_5}
\title{This is data is the miRNA2_5 FC for 2-5 time point.
This data set is used in documentation examples.}
\description{
This is data is the miRNA2_5 FC for 2-5 time point.
This data set is used in documentation examples.
}
\keyword{data}
