% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{miRNA_fc}
\alias{miRNA_fc}
\title{This is data is the combined miRNA FC for all time points.
This data set is used in documentation examples.}
\description{
This is data is the combined miRNA FC for all time points.
This data set is used in documentation examples.
}
\keyword{data}
