% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirRnaDensityCor.R
\name{mirRnaDensityCor}
\alias{mirRnaDensityCor}
\title{mirRnaDensityCor for miRTarRNASeq miRNA and mRNA correlation real data versus sampled data}
\usage{
mirRnaDensityCor(corr0, corrS, pvalue = 0.05)
}
\arguments{
\item{corr0}{data.frame results of corMirnaRna function.}

\item{corrS}{data.frame results from the sampCorRnaMirna function.}

\item{pvalue}{The p value threshold to be used on the data density plot default is 0.05.}
}
\value{
Density plot
}
\description{
This function draws density plots for miRNA and mRNA correlation while
comparing real data vs sampled data. It mainly illustrates the where the lower %5 (sig)
relationships lie.
}
\examples{
x <- mirRnaDensityCor(corr_0, outs, pvalue = 0.05)
}
\keyword{Density}
\keyword{plot}
