% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirRnaHeatmapDiff.R
\name{mirRnaHeatmapDiff}
\alias{mirRnaHeatmapDiff}
\title{mirRnaHeatmapDiff pheatmap for miRTarRNASeq miRNA and mRNA correlation}
\usage{
mirRnaHeatmapDiff(
  finalF,
  ...,
  upper_bound = 0,
  main = "Default mRNA miRNA heatmap",
  color = c("grey90", viridis::inferno(50)),
  fontsize = 7
)
}
\arguments{
\item{finalF}{data.frame results of corMirnaRnaMiranda or corMirnaRna function}

\item{...}{arguments passed onto pheatmap}

\item{upper_bound}{is the upper_bound of the correlation pheatmap scale
default is zero user can set to values based on output of correlation result (value)}

\item{main}{is the title of the pheatmap}

\item{color}{default inferno(50) from the library viridis R base,
R colorbrewer and viridis compatible}

\item{fontsize}{default is 7 user adjustable}
}
\value{
pheatmap Obj
}
\description{
This function draws pheatmaps for miRNA and mRNA correlation while
using default and pheatmap for all other parameters
}
\examples{
x <- mirRnaHeatmapDiff(results$corrs, upper_bound = -0.1, color = rainbow(50), fontsize = 10)
}
\keyword{color,}
\keyword{correlation}
\keyword{heatmap,}
\keyword{pheatmap,}
\keyword{plot,correlation_plot}
