% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_run.R
\name{modelCoefficients}
\alias{modelCoefficients}
\title{Obtain coefficients}
\usage{
modelCoefficients(x)
}
\arguments{
\item{x}{fitted model}
}
\description{
Obtain coefficients
}
\examples{
modelCoefficients(some_model)
}
