% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_run.R
\name{modelData}
\alias{modelData}
\title{Obtain model input data}
\usage{
modelData(x)
}
\arguments{
\item{x}{fitted model}
}
\value{
Input data for the fitted model
}
\description{
Obtain model input data
}
\examples{
x <- modelData(some_model)
}
