% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_run.R
\name{modelModelName}
\alias{modelModelName}
\title{Obtain model name}
\usage{
modelModelName(x)
}
\arguments{
\item{x}{fitted model}
}
\value{
model name
}
\description{
Obtain model name
}
\examples{
modelModelName(some_model)
}
