% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_run.R
\name{modelModelPvalue}
\alias{modelModelPvalue}
\title{Obtain model p-value}
\usage{
modelModelPvalue(x)
}
\arguments{
\item{x}{fitted model}
}
\description{
Obtain model p-value
}
\examples{
modelModelPvalue(some_model)
}
