% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_run.R
\name{modelTermPvalues}
\alias{modelTermPvalues}
\title{Obtain p-values for terms in model formula}
\usage{
modelTermPvalues(x)
}
\arguments{
\item{x}{fitted model}
}
\description{
Obtain p-values for terms in model formula
}
\examples{
modelTermPvalues(some_model)
}
