% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mRNA}
\alias{mRNA}
\title{This is data is the mRNA expression file.
This data set is used in documentation examples.}
\description{
This is data is the mRNA expression file.
This data set is used in documentation examples.
}
\keyword{data}
