% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mRNA0_2}
\alias{mRNA0_2}
\title{This is data is the mRNA0_2 FC for 0-2 time point.
This data set is used in documentation examples.}
\description{
This is data is the mRNA0_2 FC for 0-2 time point.
This data set is used in documentation examples.
}
\keyword{data}
