% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mRNA_fc}
\alias{mRNA_fc}
\title{This is data is the combined mRNA FC for all time points.
This data set is used in documentation examples.}
\description{
This is data is the combined mRNA FC for all time points.
This data set is used in documentation examples.
}
\keyword{data}
