% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mRNA_fc2}
\alias{mRNA_fc2}
\title{This data is the mRNA fold change data set for difference or interrelation section.
This data set is used in documentation examples.}
\description{
This data is the mRNA fold change data set for difference or interrelation section.
This data set is used in documentation examples.
}
\keyword{data}
