% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{outs}
\alias{outs}
\title{This is data is the output file resulted from time point/conditions background correlation model.
This data set is used in documentation examples.}
\description{
This is data is the output file resulted from time point/conditions background correlation model.
This data set is used in documentation examples.
}
\keyword{data}
