% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFit.R
\name{plotFit}
\alias{plotFit}
\title{Plot model}
\usage{
plotFit(model)
}
\arguments{
\item{model}{linear model}
}
\value{
does not return value
}
\description{
Plot 2D description
}
\examples{
plotFit(lm(x ~ y, data = data.frame(x = runif(10), y = runif(10))))
}
\keyword{2d,}
\keyword{model}
\keyword{plot,}
