% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotResiduals.R
\name{plotResiduals}
\alias{plotResiduals}
\title{Plot residuals}
\usage{
plotResiduals(model)
}
\arguments{
\item{model}{linear model}
}
\description{
Plot residuals description
}
\examples{
plotResiduals(lm(x ~ y, data=data.frame(x=runif(10), y=runif(10))))
}
\keyword{model}
\keyword{plot,}
\keyword{residuals,}
