% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{results}
\alias{results}
\title{This is data is the output file resulted from time point or conditions or correlation or interrelation model.
This data set is used in documentation examples.}
\description{
This is data is the output file resulted from time point or conditions or correlation or interrelation model.
This data set is used in documentation examples.
}
\keyword{data}
