% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_run.R
\name{runModel}
\alias{runModel}
\title{Run a model of a specific kind}
\usage{
runModel(x, data, ..., model = glm_gaussian())
}
\arguments{
\item{x}{model formula}

\item{data}{data.frame to run the model on}

\item{...}{passed on to \code{fit()}}

\item{model}{model type}
}
\description{
Run a model of a specific kind
}
