% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sig_corrs}
\alias{sig_corrs}
\title{This is data is the output file resulted from time point or conditions experiment for correlation model after filtering and threshold modification.
This data set is used in documentation examples.}
\description{
This is data is the output file resulted from time point or conditions experiment for correlation model after filtering and threshold modification.
This data set is used in documentation examples.
}
\keyword{data}
