% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{some_model}
\alias{some_model}
\title{This is data is the results file from regression analysis and its estimates.
This data set is used in documentation examples.}
\description{
This is data is the results file from regression analysis and its estimates.
This data set is used in documentation examples.
}
\keyword{data}
