% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshSig.R
\name{threshSig}
\alias{threshSig}
\title{threshSig Using shuffling threshold finds appropriate significant miRNA-mRNA correlation}
\usage{
threshSig(corr0, corrS, pvalue = 0.05)
}
\arguments{
\item{corrS}{vector of correlations, from the sampCorRnaMirna function.}

\item{pvalue}{The p value threshold to be used on the sampled data.}

\item{corr_0}{data.frame results of corMirnaRna function.}
}
\value{
A dataframe of Significant mRNA and miRNA
}
\description{
This function uses the sampCorRnaMirna shuffled output to determine an appropriate threshold
for significant mRNA and miRNA relationship of the dataset and shows all those with significant
relationships.
}
\examples{
\donttest{
x <- mirRnaHeatmap(outs, corr_0)
}

}
\keyword{Significance,}
\keyword{Threshold}
