% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tzTrans.R
\name{tzTrans}
\alias{tzTrans}
\title{tzTransTranspose and z-score transformation}
\usage{
tzTrans(x)
}
\arguments{
\item{x}{matrix of miRNA or mRNA or the data.frame to be transformed}
}
\value{
transposed and transformed version of x as a matrix.
}
\description{
Transposes and z-score transforms a matrix or data.frame.
}
\examples{
x <- tzTrans(miRNA)
}
\keyword{scale,}
\keyword{transform,}
\keyword{transpose}
\keyword{zscore,}
