% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_zeroinfl.R
\name{glm_zeroinfl_negbin}
\alias{glm_zeroinfl_negbin}
\title{alias for glm_zeroinfl("negbin")}
\usage{
glm_zeroinfl_negbin(...)
}
\description{
alias for glm_zeroinfl("negbin")
}
\examples{
x <- glm_zeroinfl_negbin()
}
